// src/app/shared/category.enum.ts
export enum Categories {
  Travel = 'Travel',
  Scifi = 'Sci-Fi',
  Cooking = 'Cooking',
  Romance = 'Romance',
  Autobiogrophy = 'Autobiography',
  Horror = 'Horror',
  New = 'New',
  Popular = 'Popular',
  Science = 'Science',
  Environment = 'Environment',
  Philosophy= 'Philosophy',
  Children='Children'
}